Bs/requirements.txt
# ===== requirements.txt (Python 3.11+) =====

# Core visualization and API
gradio>=5.38.0,<6.0.0
fastapi>=0.103.0,<1.0.0
uvicorn[standard]>=0.35.0

# Data models and validation
pydantic>=2.3.0,<3.0.0
python-multipart>=0.0.20

# Multimedia processing
python-ffmpeg>=2.0.12
ffmpeg-python>=0.2.0
pydub>=0.25.1
pillow>=10.4.0

# Computer vision
opencv-python>=4.12.0.88,<5.0.0

# Numerical computing and utilities
numpy>=1.24.0,<2.0.0
tqdm>=4.67.1

# System and environment
psutil>=7.0.0
python-dotenv>=1.1.1

# Packaging and distribution
PyInstaller>=6.0.0,<7.0.0
cryptography>=41.0.0